Use django-localflavor for phone number, state
Instead of using a custom field for states and django-phonenumber-field
for cell phone numbers, use the django-localflavor package (formerly
part of Django proper).

While accepting international phone numbers is certainly a plus, the
fact is that most participants have US phone numbers, and expecting them
to put country code (+1) without an explicit warning as to their
"malformed number" is unrealistic.

In the future, emergency contacts should be allowed to be international
phone numbers (e.g. international students with family members in home
country).