Add some basic controls for WSC
Add a simple "view participant" DetailView, to be accessible to leaders
and WSC only. This view has significant boilerplate and a hack to get
the data to display cleanly. A better implementation would be able
to render all parts of the Participant model cleanly.

Also includes the ability to manage participants - setting whether or
not they've attended lectures, or outright deleting them.