Add simple creation + viewing of trips
Currently, leaders can create trips where they are not one of the
leaders (this is reasonable behavior - other leaders may be travelling,
etc. and request that another create a trip for them). While the
creating leader is selected as a leader by default, they can unselect
themselves, and not be aware that they're not leading. Some warning
should exist.

This feature requires more work.