Implement visual feedback to "I have car" with JS
Modify the `update_info` view to submit the value of "I have a car"
to the template (enabling a default value of True/False). On a failed
POST, the form will regenerate, appropriately validating or clearing
the CarForm based on whether or not the participant claims to own
a car.

This solution includes JavaScript to show/hide car rows to make the
checkbox's function clear. However, it doesn't rely on JavaScript to
function.

This fixes the bug of all car fields being marked as "required" if a
user POSTs an invalid form (but selected that they don't have a car).
Additionally, if they say they don't have a car on submission, that
will be reflected on the rendered POSTed form (with errors).