Add ability to sign up for Trips
A SignUp model mediates the relationship between Participants and the
Trips they've signed up for.

When trips are available for sign up, participants can sign up, which
creates a SignUp record (ensuring that each participant only signs up
once for each trip).

The way that participants are forced to sign up only once is kind of
broken, however.

Ideally, we would just ensure that participants are making their
first signup within clean() of the SignUp model. However, we only add
the `participant` object (through `request` in a corresponding view)
_after_ the form is validated. This means that `clean()` doesn't have
access to `participant`.

The chosen solution is to hide forms when they shouldn't apply (i.e. use
templates to prevent letting a user sign up twice).

However, since this won't necessarily prevent users from accidentally
re-sending POST data (or doing it on purpose), there's some basic
handling of edge cases and potential errors, the most primitive of which
is an assertion on a `save(commit=True)` of `SignUp`. The overridden
`save()` is the "last line of defense." That is, some template/view
should cleanly prevent or handle errors long before that.

This solution is less than ideal, but after much thought, it's the most
straightforward solution (to my knowledge).