Implement initial lottery system
The lottery system is run manually, as a management command, distinct
from this project's automated workings. (The administrator could easily
set this up as a cron job, but handling edge cases and potential errors
means this is probably a poor decision).

The lottery system is fairly crude: prioritizing certain participants
over others, it assigns participants one-by-one to their top choice,
available trip. Drivers can displace existing participants if that trip
is lacking sufficient drivers.

This algorithm needs a _lot_ of work, but it gets things working.