Automatically redirect leaders to the admin page
There are two main ways of interacting with a trip - as a potential
participant viewing info (and potentially signing up!), or as a leader
administrating the trip (electing who's on the trip, viewing medical
info, et cetera). For clarity and to avoid potential data leaks, these
two functions are separated into separate views. However, a trip leader
does not need to see the standard trip info + signup form that everyone
else doees. If a trip leader is viewing the trip, a simple redirect
will bring them to its admin page.

This also solves the problem that `admin_trip/` has no links in the
current interface.

Because leaders can no longer see the `view_trip` interface, display
all signups on the admin page.