Segment trips into past and current trips
Use Python lists to only evaluate the trip queryset once, store results
in memory. This should speed up database operations, especially when
there is a large number of past trips.

We may wish to optimize this a bit later. For example, do _all_ past
trips need to be displayed? Should we paginate results? Should past
trips only be shown from this year, and not previous years?