Disable deletion of waiting list signups
If deleting a waiting list signup while deleting a trip signup, the
signals will fail to handle this appropriately, leaving an empty
spot on the trip. (This has to do with the order in which models are
saved). It doesn't make much sense for a leader to delete waiting list
signups if they're automatically handled anyway.