Email a leader when added to a trip
This serves as a mechanism for informing all co-leaders that they're on
a trip. Presumably, they'll already be aware of another leader's
intention to create a trip and mark them as a co-leader. However, this
should serve as helpful confirmation that the trip has been created, and
that they're specified as a leader. Additionally, it can correct any
mistakenly-specified co-leaders.