Empty the waitlist before deleting a trip
When deleting a `Trip`, its child `SignUp` objects must first be
deleted. Deleting the `SignUp` objects that are on the trip can
temporarily move a waitlisted signup onto the trip. This should not be a
major problem (as the trip is soon to be deleted anyway). However,
emails are sent telling participants that they made it onto the trip.

To avoid this, ensure that there are no participants in the waiting list
before deleting a `Trip` and related objects.