Fix for email being sent on every trip update
Even when the list of leaders does not change, a `post_add` signal for
the `trip.leaders` ManyToMany field will indicate that all leaders
have been added. This happens because the set is first cleared, then
each leader is added. This would result in a slew of unnecessary emails.
As a compromise to this behavior, send out an email any time the leaders
list changes from editing a trip.