Only allow US states and DC for states
Instead of allowing US territories, which _do_ issue license plates
valid in the continental US (see: 'United States license plate designs
and serial formats' on Wikipedia), just stick with 50 states and DC.
This makes for an easier interface. In the future, we should probably
allow an "international" option for any Canadian/Mexican/etc. cars).

Additionally, just use simple dropdown instead of Select2 (as the
element wasn't styled, and it's not terribly hard to find your state in
a list of 50, even on mobile devices).