Remove Student as an affiliation
Students were mistakenly being treated as MIT affiliates in the code.
Per MITOC rules, students from other institutions get discounted
membership, but shouldn't be given preference in Winter School.

Affiliation could now easily be converted to a boolean, but we may wish
to change the field later, so leave it as-is.