Break password-changing loop
`allauth` defaults to a loop in which changing your password redirects
you to the login page, which redirects you to change your password,
which redirects you...

Change this behavior so that a successful password change just prompts
you to log in (which can redirect wherever it wants, the home page in
this case). See https://github.com/pennersr/django-allauth/issues/468
for more.