Don't favor trip attendance over no trips
The existing system aims to put "flaky" participants lower in the
resolution order. However, by using the raw reliability score, we can
unintentionally give preference to those who've been on trips when
compared to those who haven't had the chance. This negates the influence
of the secondary sorting metric - number of trips.