Use local time instead of UTC in certain contexts
Despite using timezone-aware objects throughout the codebase, many dates
and datetimes were created assuming local times, not UTC times. This
creates a number of problems (for instance, the date on a UTC time can
differ from the local time's).

Where it is appropriate, use local time instead of UTC. In cases of time
diffs, there's no need to get times expressed in local time, so leave
them as-is.