Enable signup for users not on waitlist/trip
SignUp records are rarely deleted. The "All signups" section for a trip
shows everyone who has signed up for a trip and not manually deleted
their signup or had this done by a leader. If a participant has a signup
in the trip's signup set, they cannot sign up again.

However, this design presents a problem. When the lottery places a
participant on another trip, they are encouraged to sign up for the
newly-converted first-come, first-serve trips. But if they've signed up
for any of these trips, they're unable to express their interest in
re-joining.

This commit solves this issue, giving participants a single button to
express their interest in a trip and re-join.