Disallow anonymous leader feedback
Consensus seems to be against this feature - the primary benefit of
feedback is the ability to contact the supplying leader(s) for more
information. Leaders should be able to be appropriately critical or
complimentary without the need for anonymity (any harsh criticisms can
always be directed to the WSC).

Furthermore, it's not too difficult to guess who supplied the supposedly
"anonymous" feedback.