Add Leader notes, let leaders view participants
Notes may contain things such as rating contingencies, and are only
viewable to leaders and the WSC.

Additionally, this allows leaders to view participant info. This was
always the intention (leaders should be able to view past feedback,
emergency info, etc.). However, testing as an admin never revealed this
issue.