Add summary for all trips (to be given to WIMP)
Add a WSC-viewable summary of all trips, with pertinent medical
information.

The WIMP view makes a large number of SQL queries. It could be improved
upon, but this page isn't heavily viewed, and it keeps the code simpler
to leave unoptimized.

As ensuring safety lecture attendance is the responsibility of leaders
and the WIMP, include a read-only lecture attendance field in tables.