Add migrations for new pairs feature
Because production-ready version was deployed before the paired lottery
signups feature, we'll need to perform the first migration. The
"initial" migration file is the database before this change, and
migration 0002 is what's needed to accomodate the new feature.

Because we removed `wed_at_noon` from `dateutils`, the initial migration
file sets the default signup opening time to the new `wed_morning`.
However, since the default for this field is a Python function (and thus
will not set a DEFAULT value in the MySQL schema), this is not a problem.