Give active leaders priority in the lottery
Dedicated MITOC leaders should be given priority when acting as
participants. In many cases, they will need to be participants in order
to be considered for a higher rating. They may also be leading another
trip that weekend - it's important to allow them priority for picking a
geographically feasible trip.

Leaders who've not led any trips will be treated as normal participants.