Hotfix for broken admin trip editing
Formerly, a POST on the admin view would result in an error. On
get_context_data, built-in generic code was checking for the unset
self.object, raising an AttributeError. To fix this, we manually assign
self.object before calling the superclass's method.

Separately, returning the results of the GET method will produce an
error if any participants are deleted. Just redirect to the URL instead.