Allow leaders to manually add participants
Because some participants are last-minute additions, we should allow
leaders the ability to manually add participants themselves. There are
also circumstances where a participant needs to go ahead of the other
waitlisted spots (e.g. they're a driver, and the only way the trip will
work). In any case, leaders need granularity of control.

This feature allows leaders to add participants one at a time, with
varying degrees of prioritization.

It also allows simultaneous deletion of a trip spot, and the addition of
the person who'll fill said spot, even with a waiting list already
present.