Fix leader-dropping bug
When a leader deletes a Car object with the update_info form, they would
lose their leader status. This occurs because the relation to the Leader
object is dropped when the related Car object is deleted for the
participant. To solve this, delete the Car after modifying the
Participant it was once linked to.