Fix bug affecting signup of carless users
The fix for leader dropping introduced a new bug - participants who sign
up and specify that they don't have a car will not have their
information saved. To be a bit more clear, always call
participant.save(), but use a boolean to denote if we should delete a
Car object.