Allow manipulation of participants after close
While participants shouldn't be allowed to sign up for a trip after it
closes, leaders should retain the ability to modify participants after a
trip closes (for example, they may wish to get the list accurate in
order to do participant reviews).

If spaces are created on a trip on the day of the trip or later,
waitlist spots still aren't bumped. However, the leader can manually
override this for any given participant.