Use specificity to workaround FooTable hide/show
FooTable was interfering with altering the display property (whether via
jQuery or vanilla JS) within a FooTable row. The simple solution is to
simply disable hide/show controls, and display the content they
initially hide. This behavior is desirable anyway- in the case of
participant feedback, it's hidden in the table by FooTable's row
toggling anyway.

Changes the expandable feedback from using an inline style to one defined
in the CSS. A selector with higher specificity automatically hides the
control and shows the expandable content if it's within a FooTable row.