Sort signups by last modified datetime
Fixes a serious mismatch between the system's designed behavior and
its implementation. Signups were initially sorted by creation time, for
simplicity in displaying them to users. However, this meant that the
"last" participant on a trip signup (i.e. the one most vulnerable to
being booted) was actually just the participant whose signup was created
last. This had the potential to punish latecomers to the lottery (when
their signups should be treated exactly the same as others').

Instead, sort by modification time. Because participants are not allowed
to edit their signups, this will allow sorting by the most recent time
their SignUp object was modified, including:
 - placement by lottery algorithm
 - placement by signals (i.e. automatic bumping)
 - manual placement by leaders

In effect, we're relying on the fact that `signup.save()` when
`on_trip` = True is only called when they're to be placed on the trip.
There are obviously cases where this might not be true (e.g. manual
admin signup modification), but they are rare enough where this simple
sorting system should be effective.