Don't "trip or wait" signups already on trip
Although this method shouldn't be called when a signup is already on the
trip, ensure that the method exits if the signup isn't actually on the
trip. Previously, calling `trip_or_wait` on a signup where `on_trip` =
True would result in removing that signup from the trip and waitlisting
them.