Only mandate notes at signup time
This solves a bug involving trips which previously did not have notes.
When the trip has no notes, signups are allowed with empty notes as
well. However, if leaders add notes to the trip, any manipulation of the
signup formsets will cause errors (as the models fail to validate).

Instead of enforcing notes' presence at the model level, make this
a feature of the participant's signup form. This also allows leaders
the ability to leave notes blank when manually signing up any
participants, regardless of whether or not there are trip notes.