Show notes on signup page, no trip if refreshed
Because the signup processing view is only seen if participants fail to
enter notes, displaying the trip notes there is very helpful (this was
per participant request).

Additionally, the form relies on a hidden input (the trip id). If the
URL is accessed directly, give a link back to the trips so that users
aren't confused by a non-submittable form.