Show open spaces on trips
We'd been doing this for a while on the production branch, but at the
cost of one extra database query per trip. Instead of implementing this
as a property of a Trip object, add an annotation for the trip list
views.

Note that this is done in a slightly hackish way. That is, signups on
the trip are counted in the annotation, but the open spaces are
calculated in the template itself. The reason for this is that there's
currently no easy way to do conditional annotation in Django 1.7.

When 1.8 is released, there will be a few concise ways to solve this
problem. Namely, there will be support for conditional aggregates
(ticket 11305). Additionally, there will be the ability use F objects in
annotations so that one could `Count()` the values of `signup__on_trip`
(see ticket 14030). Lastly, a Case expression could be used to sum only
signups where on_trip is True (ticket 24031).

This solutions works well enough, but should probably be substituted for
one of the aforementioned solutions when 1.8 is released!