Reduce query count on trip viewing page
The primary source of excess queries came from displaying participant
information for each waitlist entry. This commit reduces the number of
queries for an average trip from ~50 to ~17.

With this commit, the number of database hits required for high-traffic
pages is satisfactory. Most pages hover around ~5-7 hits. Complex
trip-viewing pages (i.e. `view_trip` and `admin_trip`) can get a bit
higher, but the major opportunities for query reduction have been taken.

The WIMP view is currently at a very dissatisfactory ~150 commits for a
weekend with 10 trips. This isn't critical, as the page is only viewed
by one person a few times a week (and the queries are all small).
However, it would probably be a good idea to optimize this eventually.