Refactor how leaders are managed in the database
Instead of a single related Leader object that represents whether or not
the participant is also a leader, link the participant to a number of
ratings. Each rating is by MITOC activity (e.g. Winter School or
climbing) and will facilitate enabling the site to handle different
trips. If a participant has any ratings, they're a leader. If they
don't, they're a normal participant.