Make separate templatetag for participants/signups
We presiously relied on a hack - Signups and Leaders both have a
participant attribute, so we can display both in a tabular fashion if
we're careful about attribute access. This is a poor assumption, and no
longer applies given the refactoring of leaders. For clarity, make one
templatetag for displaying a list of participants, and one for signups.