Fix leader application display
An artifact of the old leader rating system in the leader application
view was causing 500's. Fix the error and patch up the form so that it's
mostly functional (it can create new ratings, but cannot currently
update an existing rating). This is the same situation as the current
individual leader form, so it's a decent patch.