Convert site menu to use Bootstrap & AngularUI
This marks the beginning of a site redesign. The new site will try to be
as close to vanilla Bootstrap (with Bootstrap components). For now,
we're going to keep using Footable (which includes custom CSS and has a
dependency on jQuery). However, we should definitely replace it with
Bootstrap built-ins when feasible.

TODO: Make a custom AngularUI build where we pull some of the default
templates that we need. Right now, that's probably just the datepicker.