Use AngularUI Select instead of Select2
AngularUI's select is slimmed-down, native Angular that plays well with
Bootstrap. It's a much better fit than the jQuery version of Select2.

We're still using Select2 in a number of places, but the goal is to
phase it out shortly.

The new leader selection dropdown is very easily customizable now. We
could supply helpful information such as an avatar, trip count, etc.
simply by modifying the template.