Safely skip the "Are you sure?" logout page
When triggering normal logouts, we kept the confirmation page for
security reasons (a maliciously embedded URL could trigger logouts for
users). Now that we can properly use CSRF tokens, make the logout a POST
straight from the menu.

For more on why logging out on GET is dangerous, see the docs:
http://django-allauth.readthedocs.org/en/stable/views.html#logout