Move JavaScript into its own file
Move the various Angular elements into proper submodules instead of the
<style> tag we were using for testing purposes.

One consequence of this is that we lose access to template and static
tags from Django. Thankfully, Djangular comes with a nifty middleware to
reverse any URL for us.

There's not currently a lot of Angular code and it's thus not too
cumbersome to keep it in one file. However, we may wish to look into
minification/compilation in the future.