Remove need for prefixes in the personal info page
Because there's a collision on the name, email, and cell phone fields of
Participant and EmergencyContact, we were using Django's prefix feature
to differentiate the POST data. However, due to what I consider a bug in
Djangular, NgForms don't properly address prefixes. So that we can keep
the forms on the same page, rename the conflicting fields so each field
on the page is unique.