Use Djangular on trip signup form
This should help alleviate confusion about the notes field, which
baffled many participants.

The view that's redirected to on a bad POST (the trip_signup template)
should really never be seen. The only way to get there is to manually
override the front-end validation and submit an empty signup where notes
are required. Nevertheless, keep the route for now.

In the future, we should just use a JSON response to create a message
warning the user that notes are required.