Remove the "manage participants" view
This is currently too unwieldy of a management system. 1,000 rows is
not acceptable to perform simple tasks such as deleting a participant or
marking them as having attended lectures.

We'll replace this later with a better system, but for now it's easiest
to just remove it (I can always handle any special WSC requests at the
console).

Leader management was also pruned - get around to deleting the template
this time.