Display leader ratings when rendering participant
For Winter School, it's really useful to display a given participant's
leader rating (for example, when the WSC needs to approve trips).

As we're only using the site for Winter School at the moment, this is an
acceptable interim change. However, we'll want to conditionally display
the _right_ rating in the right context (e.g. if it's a climbing trip,
show the leaders' climbing ratings).