Add concept of "trip order"
Display participants in the order that they were placed instead of
sorting by when the signup was last updated.

This is necessary for manual participant control - if a leader "ranks"
participants, but then sees that ranking is not applied, there'll be
confusion. Additionally, this will add some fairness when leaders are
trying to decide who to drop off an overfull trip.

TODO: Make sure the lottery implements this