Include trip level in model
The WSC rightly pointed out that it's not clear what level leaders think
a trip should be rated. Include this as a mandatory field when creating
trips.

This is also useful for other activity types. Climbing, for instance,
should clearly state whether a trip is sport/trad, single-pitch,
multi-pitch, etc.