User profile page - beta
Create a user profile page to act as the central resource for MITOC
members. Because this is still having features added to it, don't add
it to the menu just yet.

This also removes the participant lookup feature, but we'll replace that
soon enough with a better Angular widget.

Major TODO's:

- Give a link when hovering Gravatar explaining what it is and how to
  change it
- Restore ability to look up participants
- Give clear UX on how to edit personal info
- Potentially hide medical info unless requested