Implement activity-specific leader filter
Add a self-validating widget to select leaders for a given activity.

When given an activity, the widget will filter leaders to just those who
can lead the activity. Additionally, the leaders' rating is completely
searchable and will display alongside their name.
When selecting leaders for a trip, the widget sh

The backend will still complain and render an error in case bad leaders
are submitted. However, this leverages front-end form validation.
Djangular can pick up on the invalid input.