Display leader ratings by activity
Display the leader's rating for a trip's given activity.

This is the first step in moving away from displaying Winter School
ratings everywhere. Note that this (and many other aspects of the site)
aren using N+1 queries. Namely, we should prefetch leaderratings for all
leaders in the TripView. Once the rehaul is complete, go play a round of
query golf.