View participants on the leader page
Link to a leader's participant detail page instead of their email in the
leader list.

Additionally, collapse ratings when viewing the page on mobile devices.
This page is only intended to be a temporary solution, but this will
at least stop the table from going out of the viewport when viewing on
a mobile device.