Ignore all feedback older than 13 months
Participant feedback is a critical means of assessing risk and ensuring
safety for everyone involved in a trip. It's important to retain
feedback from the past in order to make sound judgments. However,
everyone grows over time - feedback older than one year might not
accurately reflect someone's present abilities.

Set the feedback retention interval to 13 months so that we can be sure
the last Winter School is always considered. All other feedback will be
not be considered in any queries.