Use vanilla Bower over django-bower
For the same reasons as outlined here:
https://github.com/dimagi/commcare-hq/pull/8917/

Additionally, allow patch versions to float with Bower's equivalent of
"twiddle wakka."

With the exception of Footable (I manually overrode some things...), we
really don't need hard versions of any of these dependencies. Change to
allow float in patch versions. Floating minor versions is probably
fine, but I'll be conservative for now.