Fix high query count on leaders and trips views
Each combination of leader and activity was generating its own query
when displaying the leader with their activity rating. On the leader
view, this was ~500 queries for 100 leaders. Trip list pages had
similarly high counts of queries.

Reduce these high queries (down to the order of ~6 per page) by
prefetching ratings in Python.