Optimize leader checks
We can check if a user is a leader by querying their (already
prefetched) groups. The is_leader() method is still generally useful
(for instance, when investigating other participants for which we
don't already have their User instances). However, when the user in
question is the requester, it's more efficient to check groups.