Bump Django version to fix aggregation bug
Django 1.8.3 merged a fix for join in how aggregation works. Previously,
when build a `Case` statement, the `When` object was creating an `INNER
JOIN`. In our case, this means that any trips without corresponding
signups would be removed from the query when attempting to annotate a
signup count.

Django 1.8.3 changes this behavior to an outer join, so we can happily
return to using `When` over raw SQL.