Make the 'level' field only apply to Winter School
For now, we'll just omit the level field from the form. In the future,
we should make a custom "new trip" page where the level section is only
rendered if the current activity is Winter School.

We should also add some better time/date widgets for signup times and a
more compact layout for trip creation.