Better UX regarding participant information
There are two key problems with the current UX regarding the collection
of personal information:

1. New users aren't aware that they must submit personal information to
   participate.
2. Existing users aren't exactly sure whey they're being redirected to
   update their personal information.

Improve the UX here by automatically redirecting new users to the "edit
profile" page. Regardless of how the user landed on the page, we will
always explain why they're there.

The "last updated" date isn't terribly useful, so move it down to the
bottom of the page.