Stop redirect to edit_profile if anonymous
If trying to access a view decorated by user_info_required (or any other
decorator that makes use of the group_required method), the user would
be redirected to edit_profile (even if they have user info).

As a remedy, just redirect back to any route that uses group_required.
Once the user logs in, we can use the decorator again to check their
group membership and redirect if needed.