Require full names from all users
Because some users are in the system already with just mononyms, we
don't want to set the model as invalid. Instead, make the user profile
form consider users without spaces in their name as invalid. When
accessing any page that requires user info, automatically redirect and
request that the user supply their full name.