Allow leaders to sign up early for trips
For certain trips (like Circus), leaders should be allowed to join the
trip instead of joining as normal participants.

Support this behavior, and create a new LeaderSignUp object to gather
and display information about their signup.

Note that this doesn't handle the case where signups are already open
and the viewer has the ability to sign up as either a leader or a
participant. In this case, we should display "Sign Up" and "Leader
Sign Up" side-by-side, with dynamic form actions created by Angular.

At this level of complexity, we should probably refactor the signup
template into a separate templatetag.