Show leader signups on trip view
When viewing a trip, participants should always be able to see the list
of leaders. When signups aren't allowed, this will be visible in the
trip description. When signups are allowed, we might want to show notes
from leaders (so show the tabl).

At this point, the signup_table has completely replaced the
participant_table, so we can remove it.