Add trip notes as signup notes placeholder
Many users were getting confused about what the notes field is for (and
not answering important questions). Others thought that the form was
"not working" when they weren't completing the lone required field.

When trip notes are present, include a placeholder to guide these users.

While we're at it, shrink the notes field - it was unnecessarily large.