Pin jQuery-UI so it doesn't require... require
jquery-ui is a dependency of jqueryui-touch-punch.

We actually can't pin jqueryui-touch-punch because of package config
problems, but we want 0.2.3 (what's currently on production).

jQueryUI 1.12 explicitly expects requireJS in ui/core.js, so we'll
get an error trying to include this version. jQueryUI 1.11 doesn't have
this problem (and the file says it will disappear in version 1.13), so
pin to 1.11 for now.