Allow fetching membership information per user
We'll soon utilize membership information in a variety of new ways:

- Showing which participants on a trip have current membership
- Sending membership renewal reminders (optional, opt-in)

For now, this basic feature grants participants much of what they were
looking for - the ability to check their own membership status.

We choose an async solution, since the geardb server is currently living
in Scripts, and can be exceptionally laggy.