Refresh Oauth2 tokens after expiration
I mistakenly assumed that oauthclient was automatically refreshing
tokens. It does not, and tokens expire after 60 minutes (so, the
application loses access to Google Sheets after 60 minutes of uptime).

Decorate the utility methods so that they can update their own tokens if
needed.

Note that (in its current state), we could have multiple Celery tasks
refreshing tokens simultaneously. This isn't so much a race condition,
since we don't really care which new refreshed token is used.
Additionally, we already use locks so that only one function can be
used at a time.

In short, this approach does not expose the possibility of race
conditions when these methods are used within a single thread/process
or when they're used with our existing Celery configuration.