Simplify & improve the membership + waiver query
Waiver status and account status are tied to a single email address, so
if you've signed a waiver recently under another email address, that's
not reported. Ideally, we'd allow this, but it's not how the gear
database currently functions (where each person has a single email, and
waivers + memberships are tied to that email).