Add section in participant view for account emails
We may wish to hide this when the user is viewing, but for now it will
prove a helpful resource to members who are confused about their
membership status as well as leaders (read: me) who are looking for a
quick way to inspect their verified email addresses without going to
the console.

Note that the UI is probably at maximum information density. We may wish
to start hiding certain things (such as the 'stats' column).