Enable submission, highlight errors before post
Improve the UX on our forms. Instead of disabling form submission when
the form is invalid, allow clicking to show a list of errors in need of
correction.

We'd previously trialed this on the "create trips" form, but with a
simple controller and manual ng-submit. Use a simple directive instead
that can handle the form controller directly (removing the need to
specify the form name and controller).