Style manage emails, add FAQ about membership
Since we're getting so many emails from confused participants, it's
probably a good idea to briefly include some commonly-asked questions.
We take this opportunity to redesign the "manage emails" page to use
Bootstrap styling instead of the default allauth setup.

Additionally, clean up the membershipStatus status by using 'data'
attribute prefixes, ng-bind, and one-way binding in the template.

Now that we've created this FAQ, we can remove the "account emails"
section from the user's profile, as the information is easily available
on the account emails page.