Require that system email be verified
Since we already track which email addresses are verified per user,
it makes sense to have a select widget for participant email addresses
where the only valid option are verified emails.

Also add in the requirement that any existing users with unverified
email addresses switch to a verified address or add their new email.

This will hopefully serve to futher the clarify the interaction of
account emails, MITOC membership emails, and participant-shared emails.