Support leaderless trips
We've always supported the creation of trips by another leader. However,
we don't really have the concept of trips where there aren't any
leaders.

This is a useful abstraction, as there are many events where it doesn't
make sense to have anyone listed as a leader. It will also help by
ceasing to pester leaders with requsets to supply feedback after a trip
in which they didn't really do any leading.

As before, participants will be able to easily email the trip creator
(in fact, this is given as a default).