Reduce confusion with the "All Participants" table
In the past, leaders and participants have expressed confusion about
what the "All Participants" table means. The table was showing everyone
who'd signed up for the trip at any point, whether it was a lottery or
first-come, first-serve trip.

However, this old paradigm confused a lot of people, who thought:

  1. "All Participants" indicated who was placed on the trip
  2. There was a problem with participants appearing in the "on trip"
     list, as well as the "all participants" list

Repurpose this list to serve two functions:

  1. Displaying everyone interested in a lottery trip
  2. Displaying everyone who was not placed on a FCFS trip (or given a
     slot in the waiting list)

Additional tweak: reorder the activity chair menu so it's the same
order as the leader actions menu.