Handle case where creator is also a participant
When writing the email selector, I hadn't considered the possibility
that a trip creator might choose to be a participant on their own trip.

Additionally, we need to change up a few aspects about trip
administration (notably, the 'contact your leaders' message, and the
possibility of seeing your own feedback in trip administration).

Change up the email selector widget as well:
- use ng-if's to get proper styling on buttons in a group
- Correct the 'showEmails' selections when selections change
  (e.g. auto-switch to some other valid group if all signups deleted,
   as we can do in the admin console)