Support active & expired leader ratings
Provides several enhancements to our leader rating system:
- Past ratings aren't overwritten when creating a new rating
- The system tracks who created a rating
- Ratings are never deleted, just deactivated
    - Leader status is checked when being marking as deactivated
- Ratings will display what the leader had at the time of the trip

Instead of deleting Winter School ratings at the end of the season, we
can simply mark the old ratings as inactive and make new ratings. This
will also allow us to stop asking for the previous season's rating, when
we can just fetch that straight from the database.