Fix display of past ratings
When mocking up a prototype, I'd goofed on the logic for calculating if
an application has a rating assigned. Since we compare the integer count
of ratings to a boolean, we weren't actually identifying any
applications for participants with more than one activity rating.

Simplify the template logic a bit too, fixing an issue where it would
falsely warn chairs that they were the only activity chair.