Start tracking lecture attendance by year
It may be useful to see which participants attended lectures for past
years. In any case, a simple boolean that we overwrite every season is
not useful, and requires manual data intervention.

For now, remove the existing form for lecture attendance. We'll enable a
new way that lets participants mark themselves as having attended.